CLI arguments refactoring
