+++ .github/workflows/check.yml
@@ -91,7 +91,7 @@ jobs:
       - name: Banner and version check
         if: ${{ always() && steps.setup_credsweeper.conclusion == 'success' }}
         env:
-          CREDSWEEPER_BANNER: "CredSweeper 1.16.1 crc32:42dc1a71"
+          CREDSWEEPER_BANNER: "CredSweeper 1.16.2 crc32:0f856906"
         run: |
           banner="$(python -m credsweeper --banner | grep CredSweeper | head -1)"
           echo "banner = '${banner}'"
