+++ credsweeper/__init__.py
@@ -24,4 +24,4 @@ __all__ = [
     "__version__"
 ]
 
-__version__ = "1.16.1"
+__version__ = "1.16.2"
