+++ credsweeper/__main__.py
@@ -3,4 +3,4 @@ import sys
 from credsweeper.main import main
 
 if __name__ == "__main__":
-    sys.exit(main())
+    sys.exit(main(sys.argv))
